Applied changes in order for the templated query to work correctly
